requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.0.3
aiofiles==23.2.1
cryptography==41.0.7
argon2-cffi==23.1.0
PyNaCl==1.5.0
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
